REFACTOR: Cleanup code & moved format functions to utils folder
